Crate sui_package_management

Enums

LockCommand

PublishedAtError

Functions

resolve_published_id

Find the published on-chain ID in the Move.lock or Move.toml file. A chain ID of None means that we will only try to resolve a published ID from the Move.toml. The published ID is resolved from the Move.toml if the Move.lock does not exist. Else, we resolve from the Move.lock, where addresses are automatically managed. If conflicting IDs are found in the Move.lock vs. Move.toml, a “Conflict” error message returns.

set_package_id

Sets the original-published-id in the Move.lock to the given id. This function provides a utility to manipulate the original-published-id during a package upgrade. For instance, we require graph resolution to resolve a 0x0 address for module names in the package to-be-upgraded, and the Move.lock value can be explicitly set to 0x0 in such cases (and reset otherwise). The function returns the existing original-published-id, if any.

update_lock_file

Update the Move.lock file with automated address management info. Expects a wallet context, the publish or upgrade command, its response. The Move.lock principally file records the published address (i.e., package ID) of a package under an environment determined by the wallet context config. See the ManagedPackage type in the lock file for a complete spec.