sui_package_management
Function set_package_id

pub fn set_package_id(
    package_path: &Path,
    install_dir: Option<PathBuf>,
    chain_id: &String,
    id: AccountAddress,
) -> Result<Option<AccountAddress>, Error>

Sets the original-published-id in the Move.lock to the given id. This function provides a utility to manipulate the original-published-id during a package upgrade. For instance, we require graph resolution to resolve a 0x0 address for module names in the package to-be-upgraded, and the Move.lock value can be explicitly set to 0x0 in such cases (and reset otherwise). The function returns the existing original-published-id, if any.