sui_package_management

Module system_package_versions

Structs

SystemPackage

SystemPackagesVersion

Constants

SYSTEM_GIT_REPO

Functions

latest_system_packages

Return the system packages snapshot for the latest known protocol version

system_packages_for_protocol

Return the latest protocol version that is not newer than the requested version (or Err if there is no such version).