Function sui_package_management::system_package_versions::system_packages_for_protocol

pub fn system_packages_for_protocol(
    version: ProtocolVersion,
) -> Result<(&'static SystemPackagesVersion, ProtocolVersion)>

Return the latest protocol version that is not newer than the requested version (or Err if there is no such version).

The returned [ProtocolVersion] is the protocol version that introduced the returned SystemPackagesVersion; this may be older than the requested version if either:

the system packages did not change when version was released, or

this binary is older than the requested version and therefore doesn’t know about the latest version of the system packages

You can distinguish these cases by comparing version with [ProtocolVersion::MAX].