sui_package_management
Function update_lock_file

pub async fn update_lock_file(
    context: &WalletContext,
    command: LockCommand,
    install_dir: Option<PathBuf>,
    lock_file: Option<PathBuf>,
    response: &SuiTransactionBlockResponse,
) -> Result<(), Error>

Update the Move.lock file with automated address management info. Expects a wallet context, the publish or upgrade command, and its response. The Move.lock file principally records the published address (i.e., package ID) of a package under an environment determined by the wallet context config. See the ManagedPackage type in the lock file for a complete spec.